# src/categories/dto/create-category.dto.ts
import { IsNotEmpty, IsString, IsOptional, IsNumber } from 'class-validator';
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class CreateCategoryDto {
  @ApiProperty({ description: 'Name of the category', example: 'Dairy' })
  @IsNotEmpty()
  @IsString()
  name: string;

  @ApiPropertyOptional({
    description: 'ID of the parent category (if any)',
    example: 1,
  })
  @IsOptional()
  @IsNumber()
  parentId?: number;
}