# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/{lang}/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Parent category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "summary": "Retrieve all categories with nested products",
        "tags": [
          "categories"
        ]
      }
    },
    "/{lang}/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Retrieve a specific category by ID with nested products",
        "tags": [
          "categories"
        ]
      },
      "put": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a specific category",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a category (soft delete)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/{lang}/categories/bulk/update": {
      "put": {
        "operationId": "CategoriesController_updateMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManyCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk update categories",
        "tags": [
          "categories"
        ]
      }
    },
    "/{lang}/categories/bulk": {
      "delete": {
        "operationId": "CategoriesController_bulkDelete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of category IDs to delete",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories deleted successfully"
          },
          "404": {
            "description": "No categories found to delete"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk delete categories (soft delete)",
        "tags": [
          "categories"
        ]
      }
    },
    "/{lang}/categories/restore/{id}": {
      "put": {
        "operationId": "CategoriesController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID to restore",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found or not deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore a soft-deleted category",
        "tags": [
          "categories"
        ]
      }
    },
    "/{lang}/categories/bulk/restore": {
      "put": {
        "operationId": "CategoriesController_bulkRestore",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of category IDs to restore",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories restored successfully"
          },
          "404": {
            "description": "No categories found to restore"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk restore soft-deleted categories",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Healthy Milk Products API",
    "description": "API for managing milk product orders",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the category",
            "example": "Dairy"
          },
          "parentId": {
            "type": "number",
            "description": "ID of the parent category (if any)",
            "example": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "description": "Optional description",
            "example": "Devices and gadgets",
            "nullable": true
          },
          "parent": {
            "description": "Parent category",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "children": {
            "description": "Child categories",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "products": {
            "description": "Products belonging to this category",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the category was soft-deleted",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "parent",
          "children",
          "products",
          "deletedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New name of the category",
            "example": "Consumer Electronics"
          },
          "description": {
            "type": "string",
            "description": "Description for the category",
            "example": "All kinds of electronic devices"
          }
        }
      },
      "UpdateManyCategoryDto": {
        "type": "object",
        "properties": {
          "updates": {
            "description": "Array of category updates",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "updates"
        ]
      },
      "BulkDeleteDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "Array of category IDs to delete",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
    }
  }
}